+++ ct/provider_test.go
@@ -2,12 +2,10 @@ package ct
 
 import (
 	"testing"
-
-	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 )
 
 func TestProvider(t *testing.T) {
-	if err := Provider().(*schema.Provider).InternalValidate(); err != nil {
+	if err := Provider().InternalValidate(); err != nil {
 		t.Fatalf("err: %s", err)
 	}
 }